requirements.txt
langchain==0.3.26
numpy==1.26.4
flask==3.1.1
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0

langchain-pinecone==0.2.8
langchain-openai==0.3.24
langchain-community==0.3.26
langchain-huggingface>=0.0.3


transformers>=4.44.0
accelerate>=0.34.0
sentencepiece>=0.2.0

-e .